:wrench: chore: rename project
