+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.antoon_app",
+    package = "fun.antoon",
 )
 
 android_resource(
     name = "res",
-    package = "com.antoon_app",
+    package = "fun.antoon",
     res = "src/main/res",
 )
