+++ android/app/src/main/java/fun/antoon/MainActivity.java
@@ -1,4 +1,4 @@
-package com.antoon_app;
+package fun.antoon;
 
 import android.os.Bundle;
