+++ android/app/src/main/java/fun/antoon/newarchitecture/modules/MainApplicationTurboModuleManagerDelegate.java
@@ -1,4 +1,4 @@
-package com.antoon_app.newarchitecture.modules;
+package fun.antoon.newarchitecture.modules;
 
 import com.facebook.jni.HybridData;
 import com.facebook.react.ReactPackage;
