+++ android/keystore.properties
@@ -1,0 +1,4 @@
+MYAPP_UPLOAD_STORE_FILE=antoon.keystore
+MYAPP_UPLOAD_KEY_ALIAS=my-key-alias
+MYAPP_UPLOAD_STORE_PASSWORD=31064878
+MYAPP_UPLOAD_KEY_PASSWORD=31064878
