+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "antoon_app",
-  "displayName": "antoon_app"
+  "name": "antoon",
+  "displayName": "antoon"
 }
